+++ src/components/home/MainContent.tsx
@@ -1,14 +1,14 @@
 import React, { useEffect, useState } from "react";
-import { Container, LargeText, ShowMoreButton} from "../styles/styledComponents";
-import { myApi } from "../axios.config";
-import { Category } from "../types/category";
+import { Container, LargeText, ShowMoreButton} from "../../styles/styledComponents";
+import { myApi } from "../../axios.config";
+import { Category } from "../../types/category";
 import { useNavigate } from "react-router-dom";
-import { Product } from "../types/product";
+import { Product } from "../../types/product";
 import BrowseTheRange from "./BrowseTheRange";
 import OurProducts from "./OurProducts";
 import { useDispatch } from "react-redux";
-import { Filtered } from "../types/filteredProduct";
-import { updateSearch } from "../redux/slices/filterSlice";
+import { Filtered } from "../../types/filteredProduct";
+import { updateSearch } from "../../redux/slices/filterSlice";
 
 
 export default function MainContent() {
