+++ src/components/home/MainHome.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Banner from "./Banner";
+import Banner from "../home/Banner";
 import MainContent from "./MainContent";
 
 export default function MainHome() {
