+++ src/components/shop/MainContentShop.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { Container } from "../styles/styledComponents";
 import Filter from "./Filter";
 import ProductsList from "./ProductsList";
 import Pagination from "./Pagination";
+import { Container } from "../../styles/styledComponents";
 
 export default function MainContentShop() {
     return (
