+++ src/components/singleproduct/ColorContainer.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
-import { ColorButton } from "../styles/styledComponents";
+import { ColorButton } from "../../styles/styledComponents";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../redux/store";
-import { add } from "../redux/slices/temporaryCart";
+import { RootState } from "../../redux/store";
+import { add } from "../../redux/slices/temporaryCart";
 
 const colors = ["#816DFA", "#000000", "#B88E2F"]
