+++ src/components/singleproduct/SizeContainer.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
-import { SizeButton } from "../styles/styledComponents";
+import { SizeButton } from "../../styles/styledComponents";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../redux/store";
-import { add } from "../redux/slices/temporaryCart";
+import { RootState } from "../../redux/store";
+import { add } from "../../redux/slices/temporaryCart";
 
 const sizes = ["L", "XL", "XS"]
