+++ src/routes/Home.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Header from "../components/Header";
-import MainHome from "../components/MainHome";
+import MainHome from "../components/home/MainHome";
 import Footer from "../components/Footer";
 
 export default function Home() {
