+++ src/routes/Shop.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Header from "../components/Header";
-import MainShop from "../components/MainShop";
+import MainShop from "../components/shop/MainShop";
 import Footer from "../components/Footer";
 
 export default function Shop() {
